Add WithDimension; update counter semantics
WithDimension clones a DataPoint with an additional dimension.  The
intended use case is to support incrementing a per-customer counter.

Counters now no longer submit if their value is zero; cumulative
counters now no longer submit if their value is the same as the last
time they submitted.  This is intended to reduce the cost of submitting
spurious data points.

Pre-existing tests are updated to pass with the new semantics; new
functionality is not yet tested.