Correct two bugs
The first had to do with the fact that dimensions are stored as
pointers; since the location of the values didn't change while cycling
in a for loop, the effect was to have multiple entries for the
last-processed item.

Also fixed timestamps to be in milliseconds, as opposed to seconds.