Replace old DataPoint with new DataPoint
Of note, this removes support for floating-point data point values.
We've determined that we do not need them in our use case; they could be
conveniently added back by creating a DataPoint interface based on a
protoDataPoint method which returns an sfx.DataPoint with the
appropriate members set.

Also of note, this version ensures that Counters, CumulativeCounters &
Gauges are goroutine-safe by hiding their metric and dimension values
from client code.  Versions of the old SetDimension/RemoveDimension
&c. were _not_ written, in the belief that dynamic adjustment of
dimensions is not something likely to be necessary.  Client code should
untrack metrics and track new ones if changes to dimensions are
required.